REFACTOR: SigninAndUp 관련 컴포넌트 PR 리뷰 반영
- 다른 파일로 분리된 타입 중 하나의 컴포넌트에서 사용하는 타입을
  컴포넌트 내부로 이동
- useCallback 제거
