+++ src/components/features/SignInAndUpForm/SignInAndUpInput.tsx
@@ -1,8 +1,17 @@
-import { useCallback, ChangeEvent } from 'react';
+import { ChangeEvent, Dispatch, SetStateAction } from 'react';
 import uuid from 'react-uuid';
 import styled from 'styled-components';
 import { absolute } from '@/styles/mixin';
 import { INPUT_INFO } from '@/utils/constants/signInAndUp';
+
+interface InputProps {
+  option: inputOptions;
+}
+interface SignInAndUpInputProps extends InputProps {
+  isValid: boolean;
+  setFormState: Dispatch<SetStateAction<FormStateProps>>;
+  [key: string]: unknown;
+}
 
 function SignInAndUpInput({
   option,
@@ -13,22 +22,28 @@ function SignInAndUpInput({
   const newId = uuid();
   const { inputType, label, placeholder, errorMsg } = INPUT_INFO[option];
 
-  const updateFormState = useCallback(
-    (e: ChangeEvent<HTMLInputElement>) => {
-      setFormState((formState: formStateProps) => ({
-        ...formState,
-        [option]: e.target.value,
-      }));
-    },
-    [option, setFormState]
-  );
+  // const updateFormState = useCallback(
+  //   (e: ChangeEvent<HTMLInputElement>) => {
+  //     setFormState((formState: FormStateProps) => ({
+  //       ...formState,
+  //       [option]: e.target.value,
+  //     }));
+  //   },
+  //   [option, setFormState]
+  // );
+  const updateFormState = (e: ChangeEvent<HTMLInputElement>) => {
+    setFormState((formState: FormStateProps) => ({
+      ...formState,
+      [option]: e.target.value,
+    }));
+  };
 
   return (
     <Wrapper>
       <label className="srOnly" htmlFor={newId}>
         {label}
       </label>
-      <Input
+      <StyledInput
         option={option}
         type={inputType}
         id={newId}
@@ -46,7 +61,7 @@ const Wrapper = styled.div`
   position: relative;
 `;
 
-const Input = styled.input<InputProps>`
+const StyledInput = styled.input<InputProps>`
   display: block;
   width: 280px;
   height: 36px;
